Upped the precision by using Point2D.Double

+++ src/me/elendrial/lightbending/objects/GraphicalObject.java
@@ -6,7 +6,7 @@ import java.awt.Graphics;
 // No real need for this let's be real. TODO: remove
 abstract public class GraphicalObject {
 	
-	public int centreX, centreY;
+	public double centreX, centreY;
 	
 	public GraphicalObject(int x, int y) {
 		centreX = x;
